refactor: refactored to make code more manageable
- added auth/basic_auth that contains basic auth functions
- added telemetry for future logging and metrics purposes
- added router to contain all the routes and register route funtions
