+++ auth/auth.go
@@ -6,10 +6,11 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
+// Authentication interface defines the behaviour for container registry and general authentication for the backend
 type Authentication interface {
 	SignUp(ctx echo.Context) error
 	SignIn(ctx echo.Context) error
-	BasicAuth(username, password string) (map[string]interface{}, error)
+	BasicAuth() echo.MiddlewareFunc
 }
 
 type auth struct {
@@ -17,6 +18,7 @@ type auth struct {
 	c     *config.RegistryConfig
 }
 
+// New is the constructor function returns an Authentication implementation
 func New(s cache.Store, c *config.RegistryConfig) Authentication {
 	a := &auth{store: s, c: c}
 	return a
